deleted redundant acts, api (since using leonard) moved fragments into a package
