+++ app/src/main/AndroidManifest.xml
@@ -26,8 +26,6 @@
             android:resource="@menu/admin_navigation_items" />
 
         <activity android:name=".LocateActivity" />
-        <activity android:name=".AdminViewActivity" />
-        <activity android:name=".ApiJsonTmp"></activity>
         <activity
             android:name=".AdminActivityFinal"
             android:launchMode="singleTask" />
@@ -40,6 +38,7 @@
         <activity
             android:name=".AdminActivity1"
             android:launchMode="singleTask" />
+        <activity android:name=".LoginActivity" />
         <activity android:name=".ChangeLanguageActivity" />
         <activity android:name=".HeatActivity" />
         <activity android:name=".PoisonActivity" />
