--- app/src/main/java/com/example/medicaldiagnosisapp/AdminViewActivity.java
@@ -1,29 +1,0 @@
-package com.example.medicaldiagnosisapp;
-
-import android.content.Intent;
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
-import android.view.View;
-import android.widget.Button;
-
-import com.google.firebase.auth.FirebaseAuth;
-
-public class AdminViewActivity extends AppCompatActivity {
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_admin_view);
-
-        //Toolbar Buttons Start Here
-
-        Button logoutButton = findViewById(R.id.logoutButton);
-        logoutButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                FirebaseAuth.getInstance().signOut();
-                Intent moveToHelp = new Intent(getApplicationContext(), HelpActivity.class);
-                startActivity(moveToHelp);
-            }
-        });
-    }
-}
