+++ app/src/main/java/com/example/medicaldiagnosisapp/LoginActivity.java
@@ -120,7 +120,7 @@ public class LoginActivity extends AppCompatActivity {
                             Toast.makeText(getApplicationContext(), "Login successful!", Toast.LENGTH_LONG).show();
                             progressBar.setVisibility(View.GONE);
 
-                            Intent intent = new Intent(LoginActivity.this, AdminViewActivity.class);
+                            Intent intent = new Intent(LoginActivity.this, AdminActivity1.class);
                             startActivity(intent);
                         }
                         else {
