+++ app/src/main/res/layout/activity_login.xml
@@ -174,4 +174,6 @@
             app:layout_constraintTop_toTopOf="parent"
             app:layout_constraintVertical_bias="0.078" />
 
+    </LinearLayout>
+
 </LinearLayout>
