+++ app/src/main/res/layout/fragment_aed.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".AedFragment">
+    tools:context=".fragments.AedFragment">
 
     <fragment
         android:name="com.google.android.gms.maps.SupportMapFragment"
