+++ app/src/main/res/layout/fragment_chas.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ChasFragment">
+    tools:context=".fragments.ChasFragment">
 
     <fragment
         android:name="com.google.android.gms.maps.SupportMapFragment"
