+++ app/src/main/res/layout/fragment_poly.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".PolyFragment">
+    tools:context=".fragments.PolyFragment">
 
     <fragment
         android:name="com.google.android.gms.maps.SupportMapFragment"
